Optimize the algorithm
This commit brings some optimizations to the algorithm which make it
faster and more reliable. Now the calculations of the bound are
skipped if the partial result is going beyond the upper bound, which
is significally faster than the previous solution.

Additionally, new methods were added to make lines of code a
meaningful name.